feat: implement guard for auth route
